Use record_as_success instead of force_success
We need the real exit_code for some command since there are decisions
based on that code, but we want to record it as success.

So, the fake exit_code is saved in the database but the real exit_code
is used in the whole flow of the building process.